refector rancher requests
