--- pkg/util/helper_test.go
@@ -1,10 +1,0 @@
-// Copyright (C) 2020, Oracle and/or its affiliates.
-// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-package util
-
-import (
-	"testing"
-)
-
-func TestSendRequest(t *testing.T) {
-}
